+++ neon_drive/__init__.py
@@ -1,3 +1,3 @@
 """Neon Drive."""
 
-__version__ = "5.4.0-beta.13"
+__version__ = "5.5.0-beta.1"